+++ R/Imports.R
@@ -4,11 +4,15 @@
 #'
 #' @import coda
 #'
-#' @importFrom graphics plot lines abline par
+#' @importFrom graphics plot lines abline par polygon
+#'
+#' @importFrom grDevices adjustcolor
 #'
 #' @importFrom stats quantile model.response model.matrix sd median rnorm na.pass
 #'
 #' @importFrom methods formalArgs
 #'
 #' @importFrom utils maintainer
+#'
+#' @importFrom zoo index
 NULL