+++ R/utilities_shrink_TVP_res.R
@@ -1,23 +1,24 @@
 
-#' Nicer printing of shrinkTVP_res objects
+#' Nicer printing of shrinkTVP objects
 #'
-#' @param x A \code{shrinkTVP_res} object
+#' @param x A \code{shrinkTVP} object
 #' @param ... Currently ignored.
 #'
 #' @return Called for its side effects and returns invisibly.
 #' @export
-print.shrinkTVP_res <- function(x, ...){
+print.shrinkTVP <- function(x, ...){
+  ind <- attr(x, "index")
   cat(paste0("Object containing a fitted TVP model ", ifelse(attr(x, "sv"), "with stochastic volatility ", ""), "with:\n",
-      " - ", formatC(length(attr(x, "colnames")), width = 7), " covariates\n",
-      " - ", formatC(length(x$model$y), width = 7), " timepoints\n",
-      " - ", formatC(attr(x, "niter"), width = 7), " MCMC draws\n",
-      " - ", formatC(attr(x, "nburn"), width = 7), " burn-n\n",
-      " - ", formatC(attr(x, "nthin"), width = 7), " thinning"))
+             " - ", formatC(length(attr(x, "colnames")), width = 7), " covariates\n",
+             " - ", formatC(length(x$model$y), width = 7), " timepoints, running from ", min(ind), " to ", max(ind), "\n",
+             " - ", formatC(attr(x, "niter"), width = 7), " MCMC draws\n",
+             " - ", formatC(attr(x, "nburn"), width = 7), " burn-n\n",
+             " - ", formatC(attr(x, "nthin"), width = 7), " thinning"))
   invisible(x)
 }
 
 #' @export
-summary.shrinkTVP_res <- function(object, digits = 3, showprior = TRUE, ...) {
+summary.shrinkTVP <- function(object, digits = 3, showprior = TRUE, ...) {
 
   # Check if digits is scalar, integer and positive
   if (is.scalar(digits) == FALSE |
@@ -33,7 +34,7 @@ summary.shrinkTVP_res <- function(object, digits = 3, showprior = TRUE, ...) {
   }
 
   ret <- attributes(object)
-  class(ret) <- c("summary.shrinkTVP_res")
+  class(ret) <- c("summary.shrinkTVP")
   ret$priorvals <- object$priorvals
   ret$summaries <- object$summaries
   ret$types <- lapply(object, function(x) return(attr(x, "type")))
@@ -42,9 +43,9 @@ summary.shrinkTVP_res <- function(object, digits = 3, showprior = TRUE, ...) {
   ret
 }
 
-#' @method print summary.shrinkTVP_res
+#' @method print summary.shrinkTVP
 #' @export
-print.summary.shrinkTVP_res <- function(x, ...) {
+print.summary.shrinkTVP <- function(x, ...) {
   cat("\nSummary of ", (x$niter - x$nburn), " MCMC draws after burn-in of ", x$nburn, ".\n", sep = "")
 
   if(x$showprior == TRUE){
@@ -147,10 +148,51 @@ print.summary.shrinkTVP_res <- function(x, ...) {
 #'
 #' \code{plot.mcmc.tvp} plots empirical posterior quantiles for a time-varying parameter.
 #'
-#' @param x a mcmc.tvp object
-#' @param probs numeric vector of quantiles to plot for each point in time, with values in [0,1].
-#' The value 0.5 is added to \code{probs} if the user does not include it.
-#' The default value is \code{c(0.025, 0.25, 0.5, 0.75, 0.975)}.
+#' @param x \code{mcmc.tvp} object
+#' @param probs vector of boundaries for credible intervals to plot for each point in time, with values in [0,1].
+#' The largest and smallest value form the outermost credible interval, the second smallest and second largest the second outermost and so forth.
+#' The default value is \code{c(0.025, 0.25, 0.75, 0.975)}. Note that there have to be the same number of probs
+#' < 0.5 as there are > 0.5.
+#' @param shaded single logical value or a vector of logical values, indicating whether or not to shade the area between the pointwise
+#' credible intervals. If a vector is given, the first value given is used to determine if the area between the outermost credible interval
+#' is shaded, the second for the second outermost and so forth. Recycled in the usual fashion if the vector is shorter than the
+#' number of quantile pairs. The default value is \code{TRUE}.
+#' @param quantlines single logical value or a vector of logical values, indicating whether or not to draw borders along the pointwise
+#' credible intervals. If a vector is given, the first value given is used to determine whether the outermost credible interval
+#' is marked by lines, the second for the second outermost and so forth. Recycled in the usual fashion if the vector is shorter than the
+#' number of credible intervals. The defualt value is \code{FALSE}.
+#' @param shadecol single character string or a vector of character strings. Determines the color of the shaded areas that represent
+#' the credible intervals. If a vector is given, the first color given is used for the outermost area,
+#' the second for the second outermost and so forth. Recycled in the usual fashion if the vector is shorter than the number
+#' of shaded areas. Has no effect if \code{shaded = FALSE}. The default value is \code{"skyblue"}.
+#' @param shadealpha real number between 0 and 1 or a vector of real numbers between 0 and 1.
+#' Determines the level of transparency of the shaded areas that represent
+#' the credible intervals. If a vector is given, the first value
+#' given is used for the outermost area, the second for the second outermost and so forth. Recycled in the usual fashion
+#' if the vector is shorter than the number of shaded areas. Has no effect if \code{shaded = FALSE}.
+#' The default value is \code{0.5}.
+#' @param quantlty either a single integer in [0,6] or one of the character strings \code{"blank",
+#' "solid", "dashed", "dotted", "dotdash", "longdash", "twodash"} or a vector containing these. Determines the line type of the borders
+#' drawn around the shaded areas that represent the credible intervals. Note that if a vector is supplied the elements have to either
+#' be all integers or all character strings. If a vector is given, the first value given is used for the outermost area, the second for
+#' the second outermost and so forth. Recycled in the usual fashion if the vector is shorter than the number of shaded areas.
+#' Has no effect if \code{quantlines = FALSE}. The default value is \code{2}.
+#' @param quantcol single character string or a vector of character strings. Determines the color of the borders drawn around the shaded
+#' areas that represent the credible intervals. If a vector is given, the first color given is used for borders of the outermost area,
+#' the second for the second outermost and so forth. Recycled in the usual fashion if the vector is shorter than the number
+#' of shaded areas. Has no effect if \code{quantlines = FALSE}. The default value is \code{"red"}.
+#' @param quantlwd single real, positive number or a vector of real, positive numbers. Determines the line width of the borders
+#' drawn around the shaded areas that represent the credible intervals. If a vector is given, the first number given is used for
+#' the borders of the outermost area, the second for the second outermost and so forth. Recycled in the usual fashion if the vector
+#' is shorter than the number of shaded areas. Has no effect if \code{quantlines = FALSE}. The default value is \code{1}.
+#' @param drawzero single logical value determining whether to draw a horizontal line at zero or not. The default value is \code{TRUE}.
+#' @param zerolty single integer in [0,6] or one of the character strings \code{"blank",
+#' "solid", "dashed", "dotted", "dotdash", "longdash", "twodash"}. Determines the line type of the horizontal line at zero. Has no effect
+#' if \code{drawzero = FALSE}. The default value is \code{2}.
+#' @param zerolwd single real, positive number. Determines the line width of the horizontal line at zero. Has no effect
+#' if \code{drawzero = FALSE}. The default value is \code{1}.
+#' @param zerocol single character string. Determines the color of the horizontal line at zero. Has no effect if \code{drawzero = FALSE}.
+#' The default value is \code{"grey"}.
 #' @param ... further arguments to be passed to \code{plot}.
 #' @return Called for its side effects and returns invisibly.
 #'
@@ -160,13 +202,19 @@ print.summary.shrinkTVP_res <- function(x, ...) {
 #' sim <- simTVP(theta = c(0.2, 0, 0), beta_mean = c(1.5, -0.3, 0))
 #' data <- sim$data
 #'
-#' output <- shrinkTVP(y ~ x1 + x2, data)
-#' plot(output$beta$beta_x1)
+#' res <- shrinkTVP(y ~ x1 + x2, data)
+#' plot(res$beta$beta_x1)
 #' }
+#' @author Peter Knaus \email{peter.knaus@@wu.ac.at}
 #' @export
-plot.mcmc.tvp <- function(x, probs = c(0.025, 0.25, 0.5, 0.75, 0.975), ...){
+plot.mcmc.tvp <- function(x, probs = c(0.025, 0.25, 0.75, 0.975),
+                          shaded = TRUE, quantlines = FALSE,
+                          shadecol = "skyblue", shadealpha = 0.5,
+                          quantlty = 2, quantcol = "black", quantlwd = 0.5,
+                          drawzero = TRUE, zerolty = 2, zerolwd = 1, zerocol = "grey", ...){
 
-  if (is.vector(probs) == FALSE | is.list(probs) == TRUE){
+  # Input checking
+  if (length(probs) == 0 || is.vector(probs) == FALSE | is.list(probs) == TRUE ){
     stop("probs has to be a vector")
   }
 
@@ -175,26 +223,98 @@ plot.mcmc.tvp <- function(x, probs = c(0.025, 0.25, 0.5, 0.75, 0.975), ...){
     stop("all elements of probs have to be numbers between >= 0 and <= 1")
   }
 
-  # Add median prob if the user did not
-  if (!0.5 %in% probs){
-    warning("Adding median to probs", immediate. = TRUE)
-    probs <- c(probs, 0.5)
+  if (length(probs[probs < 0.5]) != length(probs[probs > 0.5])){
+    stop ("There has to be an equal amount of probs above and below 0.5")
   }
+
+  if (length(quantlines) == 0 || any(sapply(quantlines, bool_input_bad))){
+    stop("all elements of quantlines have to be logical values")
+  }
+
+  if (length(shaded) == 0 || any(sapply(shaded, bool_input_bad))){
+    stop("all elements of shaded have to be logical values")
+  }
+
+  if (is.character(shadecol)){
+    if (any(sapply(shadecol, char_input_bad))){
+      stop("shadecol has to be a string or a positive number or a vector of strings or positive numbers")
+    }
+  } else if (is.numeric(shadecol)){
+    if (any(sapply(shadecol, numeric_input_bad))){
+      stop("shadecol has to be a string or a positive number or a vector of strings or positive numbers")
+    }
+  } else {
+    stop("shadecol has to be a string or a positive number or a vector of strings or positive numbers")
+  }
+
+
+  if (length(shadealpha) == 0 || any(shadealpha > 1) | any(shadealpha < 0) | any(sapply(shadealpha, numeric_input_bad_zer))){
+    stop("all elements of shadealpha have to be numbers between >= 0 and <= 1")
+  }
+
+  if (length(quantlty) == 0 || any(sapply(quantlty, lty_input_bad))) {
+    stop("every element of quantlty has to be an integer from 0 to 6 or one of 'blank', 'solid', 'dashed', 'dotted', 'dotdash', 'longdash', or 'twodash'")
+  }
+
+  if (is.character(quantcol)){
+    if (any(sapply(quantcol, char_input_bad))){
+      stop("quantcol has to be a string or a positive number or a vector of strings or positive numbers")
+    }
+  } else if (is.numeric(quantcol)){
+    if (any(sapply(quantcol, numeric_input_bad))){
+      stop("quantcol has to be a string or a positive number or a vector of strings or positive numbers")
+    }
+  } else {
+    stop("quantcol has to be a string or a positive number or a vector of strings or positive numbers")
+  }
+
+  if (length(quantlwd) == 0 || any(sapply(quantlwd, numeric_input_bad))){
+    stop("quantlwd has to be either a positive number or a string of positive numbers")
+  }
+
+  if (bool_input_bad(drawzero)){
+    stop("drawzero has to be a logical")
+  }
+
+  if (lty_input_bad(zerolty)){
+    stop("zerolty has to be an integer from 0 to 6 or one of 'blank', 'solid', 'dashed', 'dotted', 'dotdash', 'longdash', or 'twodash'")
+  }
+
+  if (numeric_input_bad(zerolwd)){
+    stop("zerolwd has to be a positive number")
+  }
+
+  if (is.character(zerocol)){
+    if (char_input_bad(zerocol)){
+      stop("zerocol has to be a string or a positive number")
+    }
+  } else if (is.numeric(zerocol)){
+    if (numeric_input_bad(zerocol)){
+      stop("zerocol has to be a string or a positive number")
+    }
+  } else {
+    stop("zerocol has to be a string or a positive number")
+  }
+
+
   # Sort by ascending size
+  probs <- c(probs, 0.5)
   probs <- sort(probs)
 
   # Calculate all quantiles
   if (length(probs) == 1){
     quants <- matrix()
   }
-  quants <- apply(x, 2, quantile, probs)
 
   # Create x_vec for plotting
-  x_vec <- 1:ncol(x)
+  x_vec <- attr(x, "index")
+
+  startpoint <- ifelse(length(x_vec) == ncol(x), 1, 2)
+  quants <- apply(x[, startpoint:ncol(x)], 2, quantile, probs)
 
   # Extract all user specified args and add standard ones that are missing
   args <- list(...)
-  standard_args <- list(ylim = c(min(quants), max(quants)), type = "l", xlab = "T", ylab = "")
+  standard_args <- list(ylim = c(min(quants), max(quants)), type = "l", xlab = "", ylab = "")
   missing_args <- names(standard_args)[!names(standard_args) %in% names(args)]
   args[missing_args] <- standard_args[missing_args]
 
@@ -203,36 +323,56 @@ plot.mcmc.tvp <- function(x, probs = c(0.025, 0.25, 0.5, 0.75, 0.975), ...){
   if (length(probs) == 1){
     args$y <- quants
   } else {
-    args$y <- quants["50%", ]
+    args$y <- quants[median(1:nrow(quants)), ]
   }
 
-
-  # Plot
+  # Create plot
   do.call(plot, args)
 
-  # Add horizontal line at 0
-  abline(h = 0, lty = 2, col = "grey")
 
-  # Plot all quantiles as red dashed lines
-  for (i in rownames(quants)){
-    if (i != "50%"){
-      lines(quants[i, ], lty = 2, col = "red")
+  # Add horizontal line at 0
+  if (drawzero){
+    abline(h = 0, lty = zerolty, col = zerocol, lwd = zerolwd)
+  }
+
+
+  nint <- ((nrow(quants) - 1)/2)
+  for (i in 1:nint){
+    currcol <- ifelse(length(shadecol) > 1, rep_len(shadecol, nint)[i], shadecol)
+    curralpha <- ifelse(length(shadealpha) > 1,  rep_len(shadealpha, nint)[i], shadealpha)
+
+    currlty <-  ifelse(length(quantlty) > 1, rep_len(quantlty, nint)[i], quantlty)
+    currbordcol <- ifelse(length(quantcol) > 1, rep_len(quantcol, nint)[i], quantcol)
+    currlwd <-  ifelse(length(quantlwd) > 1, rep_len(quantlwd, nint)[i], quantlwd)
+    currbord <- ifelse(length(quantlines) > 1, rep_len(quantlines, nint)[i], quantlines)
+
+    polygon(c(x_vec, rev(x_vec)), c(quants[i, ], rev(quants[nrow(quants) + 1 - i, ])),
+            col = adjustcolor(currcol, alpha.f = curralpha), border = FALSE)
+
+    if (currbord == TRUE){
+      lines(y = quants[i, ], x = x_vec, col = currbordcol, lwd = currlwd, lty = currlty)
+      lines(y = quants[nrow(quants) + 1 - i, ], x = x_vec, col = currbordcol, lwd = currlwd, lty = currlty)
     }
   }
+
+  # re paint line to be on top
+  do.call(lines, args[names(args) %in% c("y", "x", "lwd", "lty", "col")])
 
 }
 
 #' Graphical summary of posterior distribution
 #'
-#' \code{plot.shrinkTVP_res} generates plots visualizing the posterior distribution.
+#' \code{plot.shrinkTVP} generates plots visualizing the posterior distribution.
 #'
-#' @param x a \code{shrinkTVP_res} object.
+#' @param x a \code{shrinkTVP} object.
 #' @param pars a character vector containing the names of the parameters to be visualized.
-#' The names have to coincide with the names of the list elements of the \code{shrinkTVP_res}
+#' The names have to coincide with the names of the list elements of the \code{shrinkTVP}
 #' object. Throws an error if any element of \code{pars} does not fulfill this criterium.
 #' The default is \code{c("beta")}.
 #' @param nplot positive integer that indicates the number of tvp plots to display on a single
 #' page before a new page is generated. The default value is 3.
+#' @param mar A numerical vector of the form \code{c(bottom, left, top, right)} which gives the number of lines of margin to be
+#' specified on the four sides of the plot, as in \code{\link{par}}. The default is c(2, 4, 1, 2) + 0.1.
 #' @param ... further arguments to be passed to the respective plotting functions.
 #' @return Called for its side effects and returns invisibly.
 #' @examples
@@ -250,8 +390,9 @@ plot.mcmc.tvp <- function(x, probs = c(0.025, 0.25, 0.5, 0.75, 0.975), ...){
 #' plot(output, pars = c("beta", "hello"))
 #' }
 #'
+#' @author Peter Knaus \email{peter.knaus@@wu.ac.at}
 #' @export
-plot.shrinkTVP_res <- function(x, pars = c("beta"), nplot = 3, ...){
+plot.shrinkTVP <- function(x, pars = c("beta"), nplot = 3, mar = c(2, 4, 1, 2) + 0.1, ...){
 
 
   # Check that any pars are specified
@@ -261,7 +402,7 @@ plot.shrinkTVP_res <- function(x, pars = c("beta"), nplot = 3, ...){
 
   # Check if pars is of type "char"
   if (is.character(pars) == FALSE){
-    stop("chars has to be of type characater")
+    stop("pars has to be of type characater")
   }
 
   # Check if all specified par values are allowed
@@ -280,8 +421,9 @@ plot.shrinkTVP_res <- function(x, pars = c("beta"), nplot = 3, ...){
   }
 
 
-  # Save mfrow value so that it can be restored after plotting
+  # Save mfrow and mar value so that it can be restored after plotting
   prev_mfrow <- par()$mfrow
+  prev_mar <- par()$mar
 
 
   # Create sublist containing only parameters specified in par and copy attributes
@@ -295,19 +437,41 @@ plot.shrinkTVP_res <- function(x, pars = c("beta"), nplot = 3, ...){
     # TVP parameters will be a list, so we differentiate between TVP and non-TVP this way
     if (is.list(obj[[i]]) == TRUE){
 
-      # Plot a max of three objects per plot
+      # Plot a max of nplot objects per plot
       if (length(obj[[i]]) > nplot){
 
-        # Change mfrow
+        # Change mfrow and mar
         par(mfrow = c(nplot, 1))
+
+        currpage <- 1
+        p_left <- length(obj[[i]])
 
         # Apply plotting method in a loop
         for (j in 1:length(obj[[i]])){
 
+          if ((j-1) %% nplot == 0){
+            # The first plot on a new page
+            currmar <- c(0, 1, 1, 1) * mar
+            xaxt <- "n"
+          } else if ((j-1) %% nplot == (nplot - 1)){
+            # The last plot on a page
+            currmar <- c(1, 1, 0, 1) * mar
+            xaxt <- "s"
+          } else if (p_left < nplot & j == length(obj[[i]])){
+            # The last plot on the last page
+            currmar <- c(1, 1, 0, 1) * mar
+            xaxt <- "s"
+          } else {
+            # Any plot in between two others
+            currmar <- c(0, 1, 0, 1) * mar
+            xaxt <- "n"
+          }
+          par(mar = currmar)
+
           # Extract all user specified args
           args <- list(...)
           # Augment user specified arguments with standard ones
-          standard_args <- list(main = paste(names(obj)[i], "of", attr(obj, "colnames")[j]))
+          standard_args <- list(ylab = paste(names(obj)[i], "of", attr(obj, "colnames")[j]), xaxt = xaxt)
           missing_args <- names(standard_args)[!names(standard_args) %in% names(args)]
           args[missing_args] <- standard_args[missing_args]
 
@@ -321,6 +485,9 @@ plot.shrinkTVP_res <- function(x, pars = c("beta"), nplot = 3, ...){
           if (j %% nplot == 0 & j != length(obj[[i]])){
             readline("Hit <Return> to see next plot: ")
           }
+
+          p_left <- p_left - currpage * nplot
+          currpage <- currpage + 1
         }
 
         # Move to next series of plots if current parameter has been thoroughly plotted
@@ -331,13 +498,28 @@ plot.shrinkTVP_res <- function(x, pars = c("beta"), nplot = 3, ...){
       } else {
 
         # If there are less than nplot parameters to plot, adjust mfrow to nr of parameters
-        par(mfrow = c(length(obj[[i]]), 1))
+        par(mfrow = c(nplot, 1))
+
         for (j in 1:length(obj[[i]])){
+
+          # Adjust mar
+          if (j == 1){
+            currmar <- c(0, 1, 1, 1) * mar
+            xaxt <- "n"
+          } else if (j == length(obj[[i]])) {
+            currmar <- c(1, 1, 0, 1) * mar
+            xaxt <- "s"
+          } else {
+            currmar <- c(0, 1, 0, 1) * mar
+            xaxt <- "n"
+          }
+          par(mar = currmar)
+
 
           # Extract all user specified args
           args <- list(...)
           # Augment user specified arguments with standard ones
-          standard_args <- list(main = paste(names(obj)[i], "of", attr(obj, "colnames")[j]))
+          standard_args <- list(ylab = paste(names(obj)[i], "of", attr(obj, "colnames")[j]), xaxt = xaxt)
           missing_args <- names(standard_args)[!names(standard_args) %in% names(args)]
           args[missing_args] <- standard_args[missing_args]
 
@@ -346,7 +528,6 @@ plot.shrinkTVP_res <- function(x, pars = c("beta"), nplot = 3, ...){
 
           # Plot
           do.call(plot, args)
-
         }
 
         if (i < length(obj)){
@@ -368,7 +549,7 @@ plot.shrinkTVP_res <- function(x, pars = c("beta"), nplot = 3, ...){
   }
 
   # Reset mfrow
-  par(mfrow = prev_mfrow)
+  par(mfrow = prev_mfrow, mar = prev_mar)
 }
 
 
@@ -415,6 +596,22 @@ int_input_bad <- function(x) {
 bool_input_bad <- function(x){
   if (is.scalar(x) == TRUE){
     return(is.na(x) | is.logical(x) == FALSE)
+  } else {
+    return(TRUE)
+  }
+}
+
+char_input_bad <- function(x){
+  if (is.scalar(x) == TRUE){
+    return(is.na(x) | is.character(x) == FALSE)
+  } else {
+    return(TRUE)
+  }
+}
+
+lty_input_bad <- function(x){
+  if (is.scalar(x) == TRUE){
+    return((x %in% 0:6 | x %in% c("blank", "solid", "dashed", "dotted", "dotdash", "longdash", "twodash")) == FALSE)
   } else {
     return(TRUE)
   }