+++ src/RcppExports.cpp
@@ -6,9 +6,32 @@
 
 using namespace Rcpp;
 
+// pred_dens_mix_approx
+arma::vec pred_dens_mix_approx(arma::vec x_test, arma::vec y_test, arma::mat theta_sr, arma::mat beta_mean, arma::vec sig2_samp, bool sv, arma::vec sv_phi, arma::vec sv_mu, arma::vec sv_sigma2, arma::cube chol_C_N_inv_samp, arma::cube m_N_samp, int M, bool log);
+RcppExport SEXP _shrinkTVP_pred_dens_mix_approx(SEXP x_testSEXP, SEXP y_testSEXP, SEXP theta_srSEXP, SEXP beta_meanSEXP, SEXP sig2_sampSEXP, SEXP svSEXP, SEXP sv_phiSEXP, SEXP sv_muSEXP, SEXP sv_sigma2SEXP, SEXP chol_C_N_inv_sampSEXP, SEXP m_N_sampSEXP, SEXP MSEXP, SEXP logSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< arma::vec >::type x_test(x_testSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type y_test(y_testSEXP);
+    Rcpp::traits::input_parameter< arma::mat >::type theta_sr(theta_srSEXP);
+    Rcpp::traits::input_parameter< arma::mat >::type beta_mean(beta_meanSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type sig2_samp(sig2_sampSEXP);
+    Rcpp::traits::input_parameter< bool >::type sv(svSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type sv_phi(sv_phiSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type sv_mu(sv_muSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type sv_sigma2(sv_sigma2SEXP);
+    Rcpp::traits::input_parameter< arma::cube >::type chol_C_N_inv_samp(chol_C_N_inv_sampSEXP);
+    Rcpp::traits::input_parameter< arma::cube >::type m_N_samp(m_N_sampSEXP);
+    Rcpp::traits::input_parameter< int >::type M(MSEXP);
+    Rcpp::traits::input_parameter< bool >::type log(logSEXP);
+    rcpp_result_gen = Rcpp::wrap(pred_dens_mix_approx(x_test, y_test, theta_sr, beta_mean, sig2_samp, sv, sv_phi, sv_mu, sv_sigma2, chol_C_N_inv_samp, m_N_samp, M, log));
+    return rcpp_result_gen;
+END_RCPP
+}
 // do_shrinkTVP
-List do_shrinkTVP(arma::vec y, arma::mat x, arma::vec a0, int niter, int nburn, int nthin, double c0, double g0, double G0, double d1, double d2, double e1, double e2, bool learn_lambda2, bool learn_kappa2, double lambda2, double kappa2, bool learn_a_xi, bool learn_a_tau, double a_xi, double a_tau, double c_tuning_par_xi, double c_tuning_par_tau, double b_xi, double b_tau, double nu_xi, double nu_tau, bool display_progress, bool ret_beta_nc, bool store_burn, bool sv, double Bsigma_sv, double a0_sv, double b0_sv, double bmu, double Bmu, double y_test, arma::vec x_test, bool LPDS);
-RcppExport SEXP _shrinkTVP_do_shrinkTVP(SEXP ySEXP, SEXP xSEXP, SEXP a0SEXP, SEXP niterSEXP, SEXP nburnSEXP, SEXP nthinSEXP, SEXP c0SEXP, SEXP g0SEXP, SEXP G0SEXP, SEXP d1SEXP, SEXP d2SEXP, SEXP e1SEXP, SEXP e2SEXP, SEXP learn_lambda2SEXP, SEXP learn_kappa2SEXP, SEXP lambda2SEXP, SEXP kappa2SEXP, SEXP learn_a_xiSEXP, SEXP learn_a_tauSEXP, SEXP a_xiSEXP, SEXP a_tauSEXP, SEXP c_tuning_par_xiSEXP, SEXP c_tuning_par_tauSEXP, SEXP b_xiSEXP, SEXP b_tauSEXP, SEXP nu_xiSEXP, SEXP nu_tauSEXP, SEXP display_progressSEXP, SEXP ret_beta_ncSEXP, SEXP store_burnSEXP, SEXP svSEXP, SEXP Bsigma_svSEXP, SEXP a0_svSEXP, SEXP b0_svSEXP, SEXP bmuSEXP, SEXP BmuSEXP, SEXP y_testSEXP, SEXP x_testSEXP, SEXP LPDSSEXP) {
+List do_shrinkTVP(arma::vec y, arma::mat x, arma::vec a0, int niter, int nburn, int nthin, double c0, double g0, double G0, double d1, double d2, double e1, double e2, bool learn_lambda2, bool learn_kappa2, double lambda2, double kappa2, bool learn_a_xi, bool learn_a_tau, double a_xi, double a_tau, double c_tuning_par_xi, double c_tuning_par_tau, double b_xi, double b_tau, double nu_xi, double nu_tau, bool display_progress, bool ret_beta_nc, bool store_burn, bool sv, double Bsigma_sv, double a0_sv, double b0_sv, double bmu, double Bmu);
+RcppExport SEXP _shrinkTVP_do_shrinkTVP(SEXP ySEXP, SEXP xSEXP, SEXP a0SEXP, SEXP niterSEXP, SEXP nburnSEXP, SEXP nthinSEXP, SEXP c0SEXP, SEXP g0SEXP, SEXP G0SEXP, SEXP d1SEXP, SEXP d2SEXP, SEXP e1SEXP, SEXP e2SEXP, SEXP learn_lambda2SEXP, SEXP learn_kappa2SEXP, SEXP lambda2SEXP, SEXP kappa2SEXP, SEXP learn_a_xiSEXP, SEXP learn_a_tauSEXP, SEXP a_xiSEXP, SEXP a_tauSEXP, SEXP c_tuning_par_xiSEXP, SEXP c_tuning_par_tauSEXP, SEXP b_xiSEXP, SEXP b_tauSEXP, SEXP nu_xiSEXP, SEXP nu_tauSEXP, SEXP display_progressSEXP, SEXP ret_beta_ncSEXP, SEXP store_burnSEXP, SEXP svSEXP, SEXP Bsigma_svSEXP, SEXP a0_svSEXP, SEXP b0_svSEXP, SEXP bmuSEXP, SEXP BmuSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -48,16 +71,14 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< double >::type b0_sv(b0_svSEXP);
     Rcpp::traits::input_parameter< double >::type bmu(bmuSEXP);
     Rcpp::traits::input_parameter< double >::type Bmu(BmuSEXP);
-    Rcpp::traits::input_parameter< double >::type y_test(y_testSEXP);
-    Rcpp::traits::input_parameter< arma::vec >::type x_test(x_testSEXP);
-    Rcpp::traits::input_parameter< bool >::type LPDS(LPDSSEXP);
-    rcpp_result_gen = Rcpp::wrap(do_shrinkTVP(y, x, a0, niter, nburn, nthin, c0, g0, G0, d1, d2, e1, e2, learn_lambda2, learn_kappa2, lambda2, kappa2, learn_a_xi, learn_a_tau, a_xi, a_tau, c_tuning_par_xi, c_tuning_par_tau, b_xi, b_tau, nu_xi, nu_tau, display_progress, ret_beta_nc, store_burn, sv, Bsigma_sv, a0_sv, b0_sv, bmu, Bmu, y_test, x_test, LPDS));
+    rcpp_result_gen = Rcpp::wrap(do_shrinkTVP(y, x, a0, niter, nburn, nthin, c0, g0, G0, d1, d2, e1, e2, learn_lambda2, learn_kappa2, lambda2, kappa2, learn_a_xi, learn_a_tau, a_xi, a_tau, c_tuning_par_xi, c_tuning_par_tau, b_xi, b_tau, nu_xi, nu_tau, display_progress, ret_beta_nc, store_burn, sv, Bsigma_sv, a0_sv, b0_sv, bmu, Bmu));
     return rcpp_result_gen;
 END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
-    {"_shrinkTVP_do_shrinkTVP", (DL_FUNC) &_shrinkTVP_do_shrinkTVP, 39},
+    {"_shrinkTVP_pred_dens_mix_approx", (DL_FUNC) &_shrinkTVP_pred_dens_mix_approx, 13},
+    {"_shrinkTVP_do_shrinkTVP", (DL_FUNC) &_shrinkTVP_do_shrinkTVP, 36},
     {NULL, NULL, 0}
 };