+++ tests/testthat/test_shrinkTVP.R
@@ -8,7 +8,7 @@ test_that("shrinkTVP executes", {
   data <- data.frame(y = runif(2), x1 = rnorm(2), x2 = rnorm(2))
   niter <- 5
 
-  expect_s3_class(shrinkTVP(y ~ ., data, niter = niter, display_progress = FALSE), "shrinkTVP_res")
+  expect_s3_class(shrinkTVP(y ~ ., data, niter = niter, display_progress = FALSE), "shrinkTVP")
 
 })