+++ NEWS.md
@@ -1,3 +1,11 @@
+## Changes in Version 1.1.0
+  - Overhauled plotting functions. The default is now shaded areas instead of lines and more customisation options are available.
+  - plot.mcmc.tvp now also adds time series information to the x-axis labels, thanks to zoo::index.
+  - Changed the way LPDS are calculated, moving them from the shrinkTVP function to a dedicated LPDS function.
+  - Added the new eval_pred_dens function, which allows the user to evaluate the one-step ahead predctive density.
+  - Objects returned from the shrinkTVP function are now of class shrinkTVP and not shrinkTVP_res. 
+  - Updated the vignette.
+
 ## Changes in Version 1.0.2:
   - Added a vignette
   - Bug fixes